refactor: extract command dispatch from main.rs
Split main.rs (1,440 lines) into three focused modules:
- cli/args.rs (623 lines): CLI type definitions (Cli, Commands, subcommands)
- cli/dispatch.rs (809 lines): command dispatch + workspace loading helpers
- main.rs (27 lines): parse CLI, init tracing, call dispatch
